requirements.txt
absl-py==0.9.0
astor==0.8.1
atari-py==0.2.6
certifi==2020.6.20
cffi==1.14.0
cloudpickle==1.3.0
cycler==0.10.0
Cython==0.29.21
fasteners==0.15
future==0.18.2
gast==0.2.2
glfw==1.12.0
google-pasta==0.2.0
grpcio==1.30.0
gym==0.17.2
h5py==2.10.0
imageio==2.9.0
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.2
monotonic==1.5
-e git://github.com/brandontrabucco/morphing-agents.git#egg=morphing_agents
mujoco-py==2.0.2.10
numpy==1.19.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
pandas==1.0.5
Pillow==7.2.0
protobuf==3.12.2
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
robel==0.1.2
scipy==1.5.1
six==1.15.0
stable-baselines[mpi]==2.10.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
transforms3d==0.3.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0